Extract conversation message mapping
