+++ apps/server/src/provider/layers/conversation-services.ts
@@ -97,6 +97,15 @@ import { RelayActivityPublisher } from "../../relay/activity-publisher.ts";
 import { RepositorySettingsService } from "../../repository-settings/services/repository-settings-service.ts";
 import { isIgnorableGrokAuthNoise } from "../drivers/acp/grok-auth-noise.ts";
 import {
+  eventToContent,
+  messageContentToText,
+  orchestrationErrorText,
+  parentItemIdOfContent,
+  roleForContent,
+  shouldIncludeInTranscript,
+  transcriptToMarkdown,
+} from "../conversation-message-mapping.ts";
+import {
   buildOrchestrationTools,
   type OrchestrationSessionTools,
   type OrchestrationToolDeps,
@@ -390,299 +399,6 @@ const messageFromRecord = (record: MessageReadRecord): Message =>
     ),
     createdAt: new Date(record.createdAt),
   });
-
-/**
- * Best-effort human string for a failed orchestration tool call. The control
- * plane never throws to the agent — failures come back as
- * `{ ok: false, error }`, and this turns a typed Effect error (or anything)
- * into the `error` string. Prefers a `reason` field, then `_tag`.
- */
-const orchestrationErrorText = (err: unknown): string => {
-  if (typeof err === "object" && err !== null) {
-    const e = err as Record<string, unknown>;
-    if (typeof e.reason === "string") return e.reason;
-    if (typeof e._tag === "string") return e._tag;
-  }
-  return "Operation failed.";
-};
-
-/**
- * Flatten a persisted message's content to a single line for `read_thread`,
- * so a spawning agent can skim what a thread has done without the full
- * structured payload.
- */
-const messageContentToText = (content: MessageContent): string => {
-  switch (content._tag) {
-    case "user":
-    case "user_rich":
-    case "assistant":
-    case "thinking":
-      return content.text;
-    case "tool_use":
-      return `[tool_use: ${content.tool}]`;
-    case "tool_result":
-      return String(content.output);
-    case "error":
-      return `[error: ${content.message}]`;
-    case "subagent_summary":
-      return content.summary;
-    default:
-      return `[${content._tag}]`;
-  }
-};
-
-/**
- * Message kinds that are pure telemetry / lifecycle noise — excluded from a
- * forked transcript copy and from the exported Markdown so the handed-off
- * context reads like a conversation, not a metrics dump.
- */
-const TRANSCRIPT_SKIP_KINDS: ReadonlySet<string> = new Set([
-  "usage",
-  "context_usage",
-  "context_compaction",
-  "usage_limit",
-]);
-
-/** Trim a serialised tool payload so a transcript export stays readable. */
-const clampBlock = (value: string, max = 2000): string =>
-  value.length > max
-    ? `${value.slice(0, max)}\n… (${value.length - max} more chars truncated)`
-    : value;
-
-const stringifyUnknown = (value: unknown): string => {
-  if (typeof value === "string") return value;
-  try {
-    return JSON.stringify(value, null, 2);
-  } catch {
-    return String(value);
-  }
-};
-
-/**
- * Render a session transcript to Markdown. Used for the "Attach transcript"
- * handoff and the copy-mode fork context file. Skips telemetry rows and
- * truncates large tool payloads.
- */
-const transcriptToMarkdown = (
-  title: string,
-  messages: ReadonlyArray<Message>,
-): string => {
-  const lines: string[] = [`# Transcript — ${title}`, ""];
-  for (const m of messages) {
-    const c = m.content;
-    if (TRANSCRIPT_SKIP_KINDS.has(c._tag)) continue;
-    switch (c._tag) {
-      case "user":
-      case "user_rich":
-        lines.push("## User", "", c.text.trim(), "");
-        break;
-      case "assistant":
-        lines.push("## Assistant", "", c.text.trim(), "");
-        break;
-      case "thinking":
-        if (!c.redacted && c.text.trim().length > 0) {
-          lines.push(
-            "> _(thinking)_ " + c.text.trim().replace(/\n/g, "\n> "),
-            "",
-          );
-        }
-        break;
-      case "tool_use":
-        lines.push(
-          `### 🛠 ${c.tool}`,
-          "",
-          "```json",
-          clampBlock(stringifyUnknown(c.input)),
-          "```",
-          "",
-        );
-        break;
-      case "tool_result":
-        lines.push(
-          c.isError ? "### ⚠ Tool result (error)" : "### Tool result",
-          "",
-          "```",
-          clampBlock(stringifyUnknown(c.output)),
-          "```",
-          "",
-        );
-        break;
-      case "error":
-        lines.push(`> **Error:** ${c.message}`, "");
-        break;
-      case "interrupted":
-        lines.push("> _(interrupted by user)_", "");
-        break;
-      case "subagent_summary":
-        lines.push(`### Sub-agent ${c.agentName}`, "", c.summary.trim(), "");
-        break;
-      default:
-        break;
-    }
-  }
-  return (
-    lines
-      .join("\n")
-      .replace(/\n{3,}/g, "\n\n")
-      .trimEnd() + "\n"
-  );
-};
-
-/**
- * Pull `parentItemId` off a content payload for the dedicated SQL column.
- * The same value is also embedded in `content_json`; the column exists for
- * indexed lookups (e.g. "all rows nested under item X").
- */
-const parentItemIdOfContent = (content: MessageContent): string | null => {
-  switch (content._tag) {
-    case "assistant":
-    case "thinking":
-    case "tool_use":
-    case "tool_result":
-    case "usage":
-    case "user_question":
-    case "user_question_answer":
-      return content.parentItemId ?? null;
-    case "context_usage":
-    case "context_compaction":
-    case "usage_limit":
-      return null;
-    case "subagent_summary":
-      // The summary row IS the wrapper; it sits at the top level next to
-      // its `Agent` tool_use. No parent.
-      return null;
-    default:
-      return null;
-  }
-};
-
-const roleForContent = (content: MessageContent): MessageRole => {
-  switch (content._tag) {
-    case "user":
-    case "user_rich":
-    case "user_question_answer":
-      return "user";
-    case "assistant":
-    case "thinking":
-    case "tool_use":
-    case "subagent_summary":
-    case "user_question":
-      return "assistant";
-    case "tool_result":
-      return "tool";
-    case "error":
-    case "interrupted":
-    case "usage":
-    case "context_usage":
-    case "context_compaction":
-    case "usage_limit":
-      return "system";
-  }
-};
-
-/**
- * Translate a provider event into the persisted message payload, or `null` if
- * the event is lifecycle-only (Started / Status / Completed / Auth / Version /
- * Capabilities / PermissionRequest). Only renderable content reaches the
- * messages table — lifecycle events drive `sessions.status` instead.
- */
-const eventToContent = (event: AgentEvent): MessageContent | null => {
-  switch (event._tag) {
-    case "AssistantMessage":
-      return {
-        _tag: "assistant",
-        text: event.text,
-        parentItemId: event.parentItemId,
-      };
-    case "Thinking":
-      return {
-        _tag: "thinking",
-        itemId: event.itemId,
-        text: event.text,
-        redacted: event.redacted,
-        parentItemId: event.parentItemId,
-      };
-    case "ToolUse":
-      return {
-        _tag: "tool_use",
-        itemId: event.itemId,
-        tool: event.tool,
-        input: event.input,
-        parentItemId: event.parentItemId,
-      };
-    case "ToolResult":
-      return {
-        _tag: "tool_result",
-        itemId: event.itemId,
-        output: event.output,
-        isError: event.isError,
-        parentItemId: event.parentItemId,
-      };
-    case "SubagentSummary":
-      return {
-        _tag: "subagent_summary",
-        itemId: event.itemId,
-        agentName: event.agentName,
-        model: event.model,
-        turns: event.turns,
-        durationMs: event.durationMs,
-        summary: event.summary,
-        isError: event.isError,
-      };
-    case "UsageDelta":
-      return {
-        _tag: "usage",
-        parentItemId: event.parentItemId,
-        inputTokens: event.inputTokens,
-        outputTokens: event.outputTokens,
-        cacheReadTokens: event.cacheReadTokens,
-        cacheCreationTokens: event.cacheCreationTokens,
-        model: event.model,
-      };
-    case "ContextUsage":
-      return {
-        _tag: "context_usage",
-        providerId: event.providerId,
-        usedTokens: event.usedTokens,
-        windowTokens: event.windowTokens,
-        precision: event.precision,
-        source: event.source,
-      };
-    case "ContextCompaction":
-      return {
-        _tag: "context_compaction",
-        itemId: event.itemId,
-        providerId: event.providerId,
-        startedAt: event.startedAt,
-        durationMs: event.durationMs,
-        beforeTokens: event.beforeTokens,
-        afterTokens: event.afterTokens,
-        status: event.status,
-      };
-    case "UsageLimit":
-      return {
-        _tag: "usage_limit",
-        providerId: event.providerId,
-        label: event.label,
-        usedPercent: event.usedPercent,
-        resetsAt: event.resetsAt,
-        windowMinutes: event.windowMinutes,
-      };
-    case "Error":
-      return { _tag: "error", message: event.message };
-    case "Interrupted":
-      return { _tag: "interrupted" };
-    case "UserQuestion":
-      return {
-        _tag: "user_question",
-        itemId: event.itemId,
-        questions: event.questions,
-        parentItemId: event.parentItemId,
-      };
-    default:
-      return null;
-  }
-};
 
 /**
  * Derive a starting title from the first line of the user's prompt. Phase 3
@@ -2652,7 +2368,7 @@ export const ConversationServicesLive = Layer.effectContext(
         const transcript = rows
           .slice(0, forkIdx + 1)
           .map(messageFromRow)
-          .filter((m) => !TRANSCRIPT_SKIP_KINDS.has(m.content._tag))
+          .filter((message) => shouldIncludeInTranscript(message.content))
           .map((m) => m.content);
 
         const resumeCursor = forkMode === "resume" ? source.cursor : null;
